fix debugging output

+++ mandocdb.c
@@ -1,4 +1,4 @@
-/*	$Id$ */
+/*	$Id: mandocdb.c,v 1.162 2014/09/03 23:21:47 schwarze Exp $ */
 /*
  * Copyright (c) 2011, 2012 Kristaps Dzonsons <kristaps@bsd.lv>
  * Copyright (c) 2011, 2012, 2013, 2014 Ingo Schwarze <schwarze@openbsd.org>
@@ -1735,7 +1735,7 @@ putkeys(const struct mpage *mpage,
 		htab = &strings;
 		if (debug > 1)
 		    for (i = 0; i < mansearch_keymax; i++)
-			if (1 << i & v)
+			if ((uint64_t)1 << i & v)
 			    say(mpage->mlinks->file,
 				"Adding key %s=%*s",
 				mansearch_keynames[i], sz, cp);
